requirements.txt
python-osc>=1.8.3
python-dotenv>=1.0.0
openai>=1.3.0
requests>=2.31.0
pyttsx3>=2.90
sounddevice>=0.4.6
soundfile>=0.12.1
requests>=2.31.0
pykakasi>=2.3.0
python-romkan-ng>=0.3.0
huggingface-hub>=0.15.1